Initial version of the package
Starting with initial set of baseline settings for configuration files for TypeScript, linting and Next.js.

Adding simple tests to validate import process for linting and Next.js.

Providing README documentation to describe the purpose of the project and include examples of how to utilize the capabilities.

Defining GitHub workflows to handle typical operations of validating future changes to the code in this repository.