Add Next.JS Support
Add a new linter configuration that is specifically designed for Next.JS.
Move the browser global vars out of the base rule set since it is part of the Next.JS ruleset. The base rules are meant for Node.JS runtime, not the browser runtime.
Migrate to defineConfig to avoid deprecated function usage in typescript-eslint.
The @next/eslint-plugin-next rules have not been included in this update as they are currently broken on the flat config. They will be included in a future update.
Update file ignore list in the base config to be more complete.